Add an explicit cast to avoid a compiler warning
ifdhandler.c:410:16: warning: comparison of integers of different signs:
      'unsigned long' and 'int' [-Wsign-compare]
                        if (*Length >= CcidSlots[reader_index].nATRLength)
                            ~~~~~~~ ^  ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~


git-svn-id: svn://anonscm.debian.org/svn/pcsclite/trunk/Drivers/ccid@5515 0ce88b0d-b2fd-0310-8134-9614164e65ea

+++ src/ifdhandler.c
@@ -407,7 +407,7 @@ EXTERNAL RESPONSECODE IFDHGetCapabilities(DWORD Lun, DWORD Tag,
 			/* If Length is not zero, powerICC has been performed.
 			 * Otherwise, return NULL pointer
 			 * Buffer size is stored in *Length */
-			if (*Length >= CcidSlots[reader_index].nATRLength)
+			if ((int)*Length >= CcidSlots[reader_index].nATRLength)
 			{
 				*Length = CcidSlots[reader_index].nATRLength;
